add helper to show edit button and delete buttons for quest depending on if user is a master

+++ app/helpers/tasks_helper.rb
@@ -29,10 +29,8 @@ module TasksHelper
 
   # Shows the delete button if you own the task.
   def show_delete_button_for_task(task)
-    if params[:id]
-      if current_user == task.user
-        link_to "Delete", task, method: :delete, data: { confirm: "Are you sure?" }, class: "btn btn-secondary"
-      end
+    if params[:id] && current_user == task.user
+      link_to "Delete", task, method: :delete, data: { confirm: "Are you sure?" }, class: "btn btn-secondary"
     end
   end
